+++ Source/Scene/CentralBodySurfaceShaderSet.js
@@ -86,7 +86,7 @@ define([
                     '   color,\n' +
                     '   u_dayTextures[' + i + '],\n' +
                     '   textureCoordinates,\n' +
-                    '   u_dayTextureTexCoordsExtent[' + i + '],\n' +
+                    '   u_dayTextureTexCoordsRectangle[' + i + '],\n' +
                     '   u_dayTextureTranslationAndScale[' + i + '],\n' +
                     (applyAlpha ?      '   u_dayTextureAlpha[' + i + '],\n' : '1.0,\n') +
                     (applyBrightness ? '   u_dayTextureBrightness[' + i + '],\n' : '0.0,\n') +