+++ Specs/Scene/SingleTileImageryProviderSpec.js
@@ -5,7 +5,7 @@ defineSuite([
          'Core/jsonp',
          'Core/loadImage',
          'Core/DefaultProxy',
-         'Core/Extent',
+         'Core/Rectangle',
          'Scene/GeographicTilingScheme',
          'Scene/Imagery',
          'Scene/ImageryLayer',
@@ -18,7 +18,7 @@ defineSuite([
          jsonp,
          loadImage,
          DefaultProxy,
-         Extent,
+         Rectangle,
          GeographicTilingScheme,
          Imagery,
          ImageryLayer,
@@ -39,16 +39,16 @@ defineSuite([
 
     it('properties are gettable', function() {
         var url = 'Data/Images/Red16x16.png';
-        var extent = new Extent(0.1, 0.2, 0.3, 0.4);
+        var rectangle = new Rectangle(0.1, 0.2, 0.3, 0.4);
         var credit = 'hi';
         var provider = new SingleTileImageryProvider({
             url : url,
-            extent : extent,
+            rectangle : rectangle,
             credit : credit
         });
 
         expect(provider.url).toEqual(url);
-        expect(provider.extent).toEqual(extent);
+        expect(provider.rectangle).toEqual(rectangle);
 
         waitsFor(function() {
             return provider.ready;
@@ -56,7 +56,7 @@ defineSuite([
 
         runs(function() {
             expect(provider.tilingScheme).toBeInstanceOf(GeographicTilingScheme);
-            expect(provider.tilingScheme.extent).toEqual(extent);
+            expect(provider.tilingScheme.rectangle).toEqual(rectangle);
             expect(provider.tileWidth).toEqual(16);
             expect(provider.tileHeight).toEqual(16);
             expect(provider.maximumLevel).toEqual(0);